requirements.txt
# Core Dependencies
tensorflow==2.4.1
opencv-python==4.5.3.56
numpy==1.19.5
Pillow==8.3.2
pytesseract==0.3.8

# Deep Learning
torch==1.9.0
torchvision==0.10.0
yolov3==1.0.0
ultralytics==8.0.0

# Image Processing
scikit-image==0.18.3
imageio==2.9.0
albumentations==1.0.3

# Data Management
pandas==1.3.3
SQLAlchemy==1.4.23
alembic==1.7.3
psycopg2-binary==2.9.1

# Web Framework & API
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
aiofiles==0.7.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
httpx==0.19.0

# Email & Notifications
python-jose[cryptography]==3.3.0
emails==0.6
Jinja2==3.0.1
aiosmtplib==1.1.6

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-cov==2.12.1
pytest-mock==3.6.1
requests==2.26.0
faker==8.14.0

# Utilities
python-dotenv==0.19.0
PyYAML==5.4.1
tqdm==4.62.3
colorama==0.4.4
rich==10.12.0
click==8.0.1

# Development Tools
black==21.9b0
flake8==3.9.2
isort==5.9.3
mypy==0.910
pre-commit==2.15.0

# Monitoring & Logging
prometheus-client==0.11.0
python-json-logger==2.0.2
sentry-sdk==1.4.3

# Documentation
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
mkdocs==1.2.3
mkdocs-material==7.3.0

# GPU Support (Optional)
# Comment out if not using GPU
tensorflow-gpu==2.4.1
torch-gpu==1.9.0+cu111


